styleai/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.8.0
replicate==1.0.4
Pillow==11.1.0
motor==3.7.0
python-dotenv==1.0.1
